[feat | #37] add confirm modal
