+++ src/pages/detailPage/DetailPage.vue
@@ -1,11 +1,5 @@
 <template>
   <div>
-    <RouterLink
-      v-for="item in componentsArray"
-      :key="item.key"
-      :to="{ path: '/detail/' + item.key }"
-      >{{ item.key }} |
-    </RouterLink>
     <component :is="currentComponent" />
   </div>
 </template>
@@ -13,27 +7,22 @@
 <style scoped></style>
 
 <script setup>
-import { ref, reactive, computed, onMounted } from 'vue'
-import { useRoute, useRouter } from 'vue-router'
+import { computed } from 'vue'
+import { useRoute } from 'vue-router'
 
 import axios from 'axios'
 
-import AddTransaction from './components/AddTransaction.vue'
-import UpdateTransaction from './components/UpdateTransaction.vue'
-import ViewTransaction from './components/ViewTransaction.vue'
+import AddTransaction from './_components/AddTransaction.vue'
+import UpdateTransaction from './_components/UpdateTransaction.vue'
+import ViewTransaction from './_components/ViewTransaction.vue'
 
 const currentRoute = useRoute()
-const router = useRouter()
 
 const componentsMap = {
   add: AddTransaction,
   view: ViewTransaction,
   update: UpdateTransaction,
 }
-
-const componentsArray = computed(() =>
-  Object.entries(componentsMap).map(([key, component]) => ({ key, component })),
-)
 
 const currentComponent = computed(() => {
   return componentsMap[currentRoute.params.action]
